Move logic to shell out to Pipenv to a PipenvRunner class
